pyproject.toml
[tool.poetry]
name = "online-search-gpt"
version = "0.6.4"
description = "Online Search GPT is an autonomous agent designed for comprehensive online research on a variety of tasks."
authors = ["Assaf Elovic <assaf.elovic@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
beautifulsoup4 = ">=4.12.2"
colorama = ">=0.4.6"
duckduckgo_search = ">=4.1.1"
md2pdf = ">=1.0.1"
openai = ">=1.3.3"
playwright = ">=1.39.0"
python-dotenv = ">=1.0.0"
pyyaml = ">=6.0.1"
uvicorn = ">=0.24.0.post1"
pydantic = ">=2.5.1"
fastapi = ">=0.104.1"
python-multipart = ">=0.0.6"
markdown = ">=3.5.1"
langchain = "^0.2"
langgraph = ">=0.0.29,<0.1"
langchain_community = "^0.2"
langchain-openai = "^0.1"
langchain-google-genai = "^1"
tavily-python = ">=0.2.8"
permchain = ">=0.0.6"
arxiv = ">=2.0.0"
PyMuPDF = ">=1.23.6"
requests = ">=2.31.0"
jinja2 = ">=3.1.2"
aiofiles = ">=23.2.1"
newspaper3k = ">=0.2.8"
SQLAlchemy = ">=2.0.28"
mistune = "^3.0.2"
htmldocx = "^0.0.6"
python-docx = "^1.1.0"
lxml = { version = ">=4.9.2", extras = ["html_clean"] }
unstructured = "^0.13.0"
tiktoken = ">=0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
